refactor(core-box): extract component logic into composable hooks
Deconstructs the monolithic `CoreBox.vue` component by extracting its core functionalities into a set of dedicated composable hooks. This significantly improves separation of concerns, maintainability, and testability.

The main responsibilities have been delegated as follows:
- `useSearch`: Manages search state, results, and execution logic.
- `useKeyboard`: Handles all keyboard navigation and shortcuts.
- `useVisibility`: Manages side effects related to window visibility.
- `useClipboard`: Encapsulates clipboard reading and auto-paste functionality.
- `useChannel`: Centralizes IPC communication with the main process.
- `TagSection.vue`: The tag display area is now a separate component.

This refactoring makes the `CoreBox` component a cleaner, more declarative view layer, with complex logic neatly organized and reusable.

BREAKING CHANGE: The `search-box.ts` and `search-core.ts` files have been deleted. Their functionalities, including `search`, `execute`, and `BoxMode`, are now encapsulated within the new hooks architecture. Any modules that depended on the old implementation must be updated.
